Improved on code and UI from previous merge

+++ Assets/Scripts/Curses/ICurseProvider.cs
@@ -5,6 +5,6 @@ namespace Game.Curses
 {
     public interface ICurseProvider
     {
-        IEnumerable<int> GetCurseModifiers(CurseEffectTypes curseEffectType);
+        IEnumerable<float> GetCurseModifiers(CurseEffectTypes curseEffectType);
     }
 }
